[Refactor]디렉터리 이름 겹쳐서 변경
