+++ src/main/java/verbly/spring/domain/quizreview/enums/ReviewQuestionType.java
@@ -1,0 +1,5 @@
+package verbly.spring.domain.quizreview.enums;
+
+public enum ReviewQuestionType {
+    cloze, mcq
+}
