+++ src/main/java/verbly/spring/domain/quizreview/enums/ReviewSessionStatus.java
@@ -1,0 +1,5 @@
+package verbly.spring.domain.quizreview.enums;
+
+public enum ReviewSessionStatus {
+    IN_PROGRESS, COMPLETED, QUIT
+}
