+++ src/main/java/verbly/spring/domain/quizreview/enums/ReviewTaskStatus.java
@@ -1,0 +1,5 @@
+package verbly.spring.domain.quizreview.enums;
+
+public enum ReviewTaskStatus {
+    PENDING, IN_PROGRESS, COMPLETED
+}
