+++ src/main/java/verbly/spring/domain/quizreview/exception/ReviewHandler.java
@@ -1,0 +1,10 @@
+package verbly.spring.domain.quizreview.exception;
+
+import verbly.spring.global.common.code.ErrorStatus;
+import verbly.spring.global.common.exception.BaseException;
+
+public class ReviewHandler extends BaseException {
+    public ReviewHandler(ErrorStatus status) {
+        super(status);
+    }
+}
