+++ src/main/java/verbly/spring/domain/quizreview/repository/ReviewTaskRepository.java
@@ -1,0 +1,23 @@
+package verbly.spring.domain.quizreview.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import verbly.spring.domain.quizreview.entity.ReviewTask;
+import verbly.spring.domain.quizreview.enums.ReviewTaskStatus;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface ReviewTaskRepository extends JpaRepository<ReviewTask, Long> {
+
+    boolean existsByUserIdAndLibraryItemId(Long userId, Long libraryItemId);
+
+    List<ReviewTask> findAllByUserId(Long userId);
+
+    List<ReviewTask> findAllByUserIdAndStatusOrderByCreatedAtAscIdAsc(Long userId, ReviewTaskStatus status);
+
+    List<ReviewTask> findAllBySessionIdOrderByTaskOrderAsc(Long sessionId);
+
+    List<ReviewTask> findAllBySessionId(Long sessionId);
+
+    Optional<ReviewTask> findBySessionIdAndTaskOrder(Long sessionId, Integer taskOrder);
+}
