+++ src/main/java/verbly/spring/domain/quizreview/service/ReviewQueryService.java
@@ -1,0 +1,8 @@
+package verbly.spring.domain.quizreview.service;
+
+import verbly.spring.domain.quizreview.dto.response.ReviewResponseDTO;
+
+public interface ReviewQueryService {
+
+    ReviewResponseDTO.QuizResultResponse getResult(Long userId, Long sessionId);
+}
