--- src/main/java/verbly/spring/domain/review/entity/ReviewAnswer.java
@@ -1,70 +1,0 @@
-package verbly.spring.domain.review.entity;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.AccessLevel;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.type.SqlTypes;
-import org.hibernate.annotations.JdbcTypeCode;
-
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(
-        name = "review_answers",
-        uniqueConstraints = {
-                @UniqueConstraint(name = "uq_review_attempt", columnNames = {"review_question_id", "attempt_no"})
-        }
-)
-public class ReviewAnswer {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    /**
-     * FK -> review_questions.id
-     */
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "review_question_id", nullable = false)
-    private ReviewQuestion reviewQuestion;
-
-    @Column(name = "attempt_no", nullable = false)
-    private int attemptNo = 1;
-
-    @JdbcTypeCode(SqlTypes.JSON)
-    @Column(name = "user_answer_json", nullable = false, columnDefinition = "json")
-    private JsonNode userAnswerJson;
-
-    @Column(name = "is_correct", nullable = false)
-    private boolean correct;
-
-    /**
-     * 오답노트
-     */
-    @Lob
-    @Column(name = "mistake_note", columnDefinition = "TEXT")
-    private String mistakeNote;
-
-    @CreationTimestamp
-    @Column(name = "answered_at", nullable = false, columnDefinition = "datetime(3)")
-    private LocalDateTime answeredAt;
-
-    public static ReviewAnswer of(ReviewQuestion q, int attemptNo, JsonNode userAnswerJson, boolean correct, String mistakeNote) {
-        ReviewAnswer a = new ReviewAnswer();
-        a.reviewQuestion = q;
-        a.attemptNo = attemptNo;
-        a.userAnswerJson = userAnswerJson;
-        a.correct = correct;
-        a.mistakeNote = mistakeNote;
-        return a;
-    }
-
-    public void updateMistakeNote(String note) {
-        this.mistakeNote = note;
-    }
-}
