--- src/main/java/verbly/spring/domain/review/enums/ReviewQuestionType.java
@@ -1,5 +1,0 @@
-package verbly.spring.domain.review.enums;
-
-public enum ReviewQuestionType {
-    cloze, mcq
-}
