--- src/main/java/verbly/spring/domain/review/enums/ReviewSessionStatus.java
@@ -1,5 +1,0 @@
-package verbly.spring.domain.review.enums;
-
-public enum ReviewSessionStatus {
-    IN_PROGRESS, COMPLETED, QUIT
-}
