--- src/main/java/verbly/spring/domain/review/enums/ReviewTaskStatus.java
@@ -1,5 +1,0 @@
-package verbly.spring.domain.review.enums;
-
-public enum ReviewTaskStatus {
-    PENDING, IN_PROGRESS, COMPLETED
-}
