--- src/main/java/verbly/spring/domain/review/service/ReviewQueryService.java
@@ -1,8 +1,0 @@
-package verbly.spring.domain.review.service;
-
-import verbly.spring.domain.review.dto.response.ReviewResponseDTO;
-
-public interface ReviewQueryService {
-
-    ReviewResponseDTO.QuizResultResponse getResult(Long userId, Long sessionId);
-}
